update to bulma 5.1 / move links to external css and js to entropy.conf.php
Moving the links to config constants allows to change their location,
e.g. use local versions which allow to run Entropy on a closed network
or a standalone pc. These constants can be set in a installer config
file.
All(?) css/js config constants are created, but not yet called by the
code (datatables, datatables-bulma), and still older css and js are
present that needs to be removed when furhter integrating bulma stuff
(issue #7) and  restructuring the html-header (part of meta-issue #31)